+++ internal/store/storage.go
@@ -13,8 +13,7 @@ var (
 type Storage struct {
 	Posts interface {
 		Create(context.Context, *Post) error
-		// TODO: mudar para int64
-		GetByID(context.Context, int) (*Post, error)
+		GetByID(context.Context, int64) (*Post, error)
 	}
 	Users interface {
 		Create(context.Context, *User) error
